+++ src/lib/useReactDataTableState/useReactDataTableState.ts
@@ -14,11 +14,18 @@ import { OptionalNullable } from "../types/NullableTypes";
 const useReactDataTableState = <TData, TFilter extends FilterModel = Record<string, never>>(
   props: OptionalNullable<useReactDataTableStateProps<TData, TFilter>>,
 ): useReactDataTableStateResult<TData, TFilter> => {
-  const { initialColumnFilters, initialSorting, initialPagination, initialRowSelection, initialExpanded, initialColumnPinning } =
-    props as useReactDataTableStateProps<TData, TFilter>;
+  const {
+    initialColumnFilters,
+    initialSorting,
+    initialPagination,
+    initialRowSelection,
+    initialExpanded,
+    initialColumnPinning,
+    initialAfterSearchFilter,
+  } = props as useReactDataTableStateProps<TData, TFilter>;
 
   const [columnFilters, setColumnFilters] = useState<TFilter>((initialColumnFilters ?? {}) as TFilter);
-  const [afterSearchFilter, setAfterSearchFilter] = useState<TFilter>((initialColumnFilters ?? {}) as TFilter);
+  const [afterSearchFilter, setAfterSearchFilter] = useState<TFilter>((initialAfterSearchFilter ?? initialColumnFilters ?? {}) as TFilter);
   const [sorting, setSorting] = useState<SortingState<TData> | undefined>(initialSorting);
   const [rowSelection, setRowSelection] = useState<RowSelectionState>(initialRowSelection ?? ({} as RowSelectionState));
   const [expanded, setExpanded] = useState<ExpandedState>(initialExpanded ?? ({} as ExpandedState));