+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- added the `usePersistentReactDataTableState` hook which saves it's own state into the local storage
+
 ### Changed
 
 - migrated to eslint 9 flat config